Removes NiceUrl methods, renames NiceUrl test cases
