+++ src/Umbraco.Web/Editors/ContentController.cs
@@ -343,7 +343,7 @@ namespace Umbraco.Web.Editors
         /// <returns></returns>
         public HttpResponseMessage GetNiceUrl(int id)
         {
-            var url = Umbraco.NiceUrl(id);
+            var url = Umbraco.Url(id);
             var response = Request.CreateResponse(HttpStatusCode.OK);
             response.Content = new StringContent(url, Encoding.UTF8, "application/json");
             return response;
