+++ src/Umbraco.Web/UmbracoHelper.cs
@@ -331,19 +331,9 @@ namespace Umbraco.Web
 
         #endregion
 
-        #region NiceUrls
+        #region Urls
 
-        /// <summary>
-        /// Returns a string with a friendly url from a node.
-        /// IE.: Instead of having /482 (id) as an url, you can have
-        /// /screenshots/developer/macros (spoken url)
-        /// </summary>
-        /// <param name="nodeId">Identifier for the node that should be returned</param>
-        /// <returns>String with a friendly url from a node</returns>
-        public string NiceUrl(int nodeId)
-        {
-            return Url(nodeId);
-        }
+        //TODO: We will need an optional culture parameter, by default it will be the current thread culture
 
         /// <summary>
         /// Gets the url of a content identified by its identifier.
@@ -366,15 +356,7 @@ namespace Umbraco.Web
             return UrlProvider.GetUrl(contentId, mode);
         }
 
-        /// <summary>
-        /// This method will always add the domain to the path if the hostnames are set up correctly.
-        /// </summary>
-        /// <param name="nodeId">Identifier for the node that should be returned</param>
-        /// <returns>String with a friendly url with full domain from a node</returns>
-        public string NiceUrlWithDomain(int nodeId)
-        {
-            return UrlAbsolute(nodeId);
-        }
+        //TODO: We will need an optional culture parameter, by default it will be the current thread culture
 
         /// <summary>
         /// Gets the absolute url of a content identified by its identifier.
